Fix #241 - Alert & Action Sheet 애니메이션 동작
- if 위치 변경
